// lib/api/schema.d.ts
export interface components {
    schemas: {
        /** ProductResponse */
        ProductResponse: {
            /** Id */
            id: number;
            /** Name */
            name: string;
            /** Description */
            description: string | null;
            /** Price */
            price: string;
            /** Category Id */
            category_id: number | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
        };
        /** ProductCreate */
        ProductCreate: {
            /** Name */
            name: string;
            /**
             * Description
             * @default null
             */
            description: string | null;
            /** Price */
            price: string;
            /**
             * Category Id
             * @default null
             */
            category_id: number | null;
        };
        /** ProductUpdate */
        ProductUpdate: {
            /**
             * Name
             * @default null
             */
            name: string | null;
            /**
             * Description
             * @default null
             */
            description: string | null;
            /**
             * Price
             * @default null
             */
            price: string | null;
            /**
             * Category Id
             * @default null
             */
            category_id: number | null;
        };
        /** CategoryResponse */
        CategoryResponse: {
            /** Id */
            id: number;
            /** Name */
            name: string;
            /** Description */
            description: string | null;
        };
        /** CategoryCreate */
        CategoryCreate: {
            /** Name */
            name: string;
            /**
             * Description
             * @default null
             */
            description: string | null;
        };
        /** CategoryUpdate */
        CategoryUpdate: {
            /**
             * Name
             * @default null
             */
            name: string | null;
            /**
             * Description
             * @default null
             */
            description: string | null;
        };
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components["schemas"]["ValidationError"][];
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}